feat: adaptive batch downloading based on buffer
This introduces adaptive batching for downloads, most useful when we
first start streaming a file. When we first start a stream, our buffer
is in an unhealthy state. We want to download the files that are next up
for the user but we don't want bandwidth competition. This downloader
will adapt to the state of the buffer:
- Critical: prioritise the file at the start of the buffer.
- Balanced: prioritise the next 2-3 files.
- Parallel: buffer healthy, maximise throughput.